+++ apps/web/src/app/components/drawer/variants/default.tsx
@@ -8,11 +8,19 @@ export const DrawerDefault = () => {
   return (
     <div>
       <Button onClick={onOpen}>Click Me!</Button>
-      <Drawer
-        isOpen={isOpen}
-        onOpenChange={onClose}
-      >
-        HIi i am drawer
+      <Drawer isOpen={isOpen} className="max-w-sm p-8" onOpenChange={onClose}>
+        <p className="mb-2 text-gray-600">
+          NATURAL AND RECYCLED MATERIALS R-LENO - Recycled wool Soft,
+          comfortable and light Designed to last a long time Resistant and
+          easily washable materials Waterproof To accompany you everywhere even
+          in case of small showers Sole - Natural and recycled rubber Soft,
+          non-slip and comfortable Insole - Removable Removable ergonomic
+          Drytech™
+        </p>
+        <p className="mb-2 text-gray-600">
+          This one specifically is the most simplest setup you can have, just a
+          simple drawer with a trigger.
+        </p>
       </Drawer>
     </div>
   );